added thread pool and made sure tasks are not calculated twice

+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/autocrafting/preview/CraftingPreviewElementError.java
@@ -31,12 +31,12 @@ public class CraftingPreviewElementError implements ICraftingPreviewElement<Item
     }
 
     @Override
-    public int getAvailable() {
+    public long getAvailable() {
         return 0;
     }
 
     @Override
-    public int getToCraft() {
+    public long getToCraft() {
         return 0;
     }
