+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/autocrafting/preview/CraftingPreviewElementItemStack.java
@@ -18,16 +18,16 @@ public class CraftingPreviewElementItemStack implements ICraftingPreviewElement<
     public static final String ID = "item_renderer";
 
     private ItemStack stack;
-    private int available;
+    private long available;
     private boolean missing;
     // If missing is true then toCraft is the missing amount
-    private int toCraft;
+    private long toCraft;
 
     public CraftingPreviewElementItemStack(ItemStack stack) {
         this.stack = ItemHandlerHelper.copyStackWithSize(stack, 1);
     }
 
-    public CraftingPreviewElementItemStack(ItemStack stack, int available, boolean missing, int toCraft) {
+    public CraftingPreviewElementItemStack(ItemStack stack, long available, boolean missing, long toCraft) {
         this.stack = stack;
         this.available = available;
         this.missing = missing;
@@ -39,18 +39,18 @@ public class CraftingPreviewElementItemStack implements ICraftingPreviewElement<
         buf.writeInt(Item.getIdFromItem(stack.getItem()));
         buf.writeInt(stack.getMetadata());
         ByteBufUtils.writeTag(buf, stack.getTagCompound());
-        buf.writeInt(available);
+        buf.writeLong(available);
         buf.writeBoolean(missing);
-        buf.writeInt(toCraft);
+        buf.writeLong(toCraft);
     }
 
     public static CraftingPreviewElementItemStack fromByteBuf(ByteBuf buf) {
         Item item = Item.getItemById(buf.readInt());
         int meta = buf.readInt();
         NBTTagCompound tag = ByteBufUtils.readTag(buf);
-        int available = buf.readInt();
+        long available = buf.readLong();
         boolean missing = buf.readBoolean();
-        int toCraft = buf.readInt();
+        long toCraft = buf.readLong();
 
         ItemStack stack = new ItemStack(item, 1, meta);
         stack.setTagCompound(tag);
@@ -96,21 +96,21 @@ public class CraftingPreviewElementItemStack implements ICraftingPreviewElement<
         GlStateManager.popMatrix();
     }
 
-    public void addAvailable(int amount) {
+    public void addAvailable(long amount) {
         this.available += amount;
     }
 
     @Override
-    public int getAvailable() {
+    public long getAvailable() {
         return available;
     }
 
-    public void addToCraft(int amount) {
+    public void addToCraft(long amount) {
         this.toCraft += amount;
     }
 
     @Override
-    public int getToCraft() {
+    public long getToCraft() {
         return this.toCraft;
     }
